add component folder, remove summary file, add mapping
