+++ code/src/App.js
@@ -1,9 +1,11 @@
 import React from 'react'
-import Header from './Header'
-import Form from './Form'
-import Footer from './Footer'
+import Header from './Components/Header'
+import Form from './Components/Form'
+import Footer from './Components/Footer'
+
 
 export const App = () => {
+  
   return (
     <>
     <div>
@@ -15,6 +17,7 @@ export const App = () => {
     <div>
       <Footer/>
     </div>
+   
     </>
   )
 }
