+++ code/src/index.css
@@ -5,8 +5,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image:url('/roses-background.jpg');
-  
+  background-image: url("/roses-background.jpg");
 }
 
 code {
@@ -16,7 +15,7 @@ code {
 
 h1 {
   font-size: 40px;
-  top: 5px;
+  margin-top: 100px;
 }
 
 .header-container {
@@ -25,58 +24,58 @@ h1 {
   height: 350px;
   /*border: 2px solid darkslategray;*/
   text-align: center;
-  background:darkslategray;
+  background: darkslategray;
   color: white;
+}
+
+.circle-1 {
+  height: 40px;
+  width: 300px;
+  background: rgb(0, 38, 255);
+  border-radius: 100%;
 }
 
 .circle {
   position: absolute;
-  top: 85px;               
-  right: 220px;
+  top: 30px;
+  right: 250px;
   height: 70px;
   width: 70px;
   background: cyan;
   border-radius: 100%;
-  
 }
 
 .last-circle {
   position: absolute;
-  top: 85px;
-  left: 220px;
+  top: 30px;
+  left: 250px;
   height: 70px;
   width: 70px;
   background: cyan;
   border-radius: 100%;
-  
 }
 
 .big-circle {
   position: absolute;
   height: 150px;
   width: 150px;
-  background:lightpink;
+  background: lightpink;
   border-radius: 100%;
-  top: 155px;
+  top: 175px;
   right: 20px;
- 
-  
 }
 
 .another-circle {
   position: absolute;
   height: 150px;
   width: 150px;
-  background:lightpink;
+  background: lightpink;
   border-radius: 100%;
-  top: 155px;
+  top: 175px;
   left: 20px;
- 
-  
 }
 
 .form {
-  
   border: 2px solid darkslategray;
   display: flex;
   justify-content: space-between;
@@ -86,8 +85,9 @@ h1 {
   text-align: center;
   background: lightpink;
   height: 500px;
+  width: 400px;
   border: 1px solid lightslategray;
-  
+
   margin: 30px;
 }
 
@@ -95,18 +95,17 @@ h1 {
   display: flex;
   align-items: center;
   justify-content: center;
-  align-self: center;  
+  align-self: center;
 }
 
 .time-container {
   padding: 10px;
   margin: 10px;
- 
+
   display: flex;
   flex-direction: row;
- justify-content: center;
- align-items: center;
-
+  justify-content: center;
+  align-items: center;
 }
 
 .genre-container {
@@ -133,7 +132,6 @@ h1 {
   border: 2px solid black;
   border-radius: 10px;
   background: cyan;
- 
 }
 
 .button-container {
@@ -148,7 +146,6 @@ h1 {
   height: 220px;
   display: flex;
   flex-direction: column;
-  
 }
 
 .footer-text {
@@ -158,4 +155,3 @@ h1 {
   margin-top: 80px;
   font-size: 20px;
 }
-
